# pkg/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice template for Kubernetes.",
        "title": "Podinfo API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/Summers1971/podinfo"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/Summers1971/podinfo/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "host": "localhost:9898",
    "basePath": "/",
    "paths": {
...
        "/cache/{key}": {
            "get": {
                "description": "returns the content from cache if key exists",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Get payload from cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to load from cache",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "writes the posted content in cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Save payload in cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to save to",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            },
            "delete": {
                "description": "deletes the key and its value from cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Delete payload from cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Key to delete",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
...
    },
...
}